# PWA-644: Use GET for GraphQL queries (but not mutations)