--- a/packages/pwa-buildpack/lib/WebpackTools/__tests__/__fixtures__/queries/getProductDetail.graphql
+++ b/packages/pwa-buildpack/lib/WebpackTools/__tests__/__fixtures__/queries/getProductDetail.graphql
@@ -1,4 +1,4 @@
-query productDetail($urlKey: String, $onServer: Boolean!) {
+query productDetail($urlKey: String) {
     productDetail: products(filter: { url_key: { eq: $urlKey } }) {
         items {
             __typename
@@ -19,9 +19,6 @@ query productDetail($urlKey: String, $onServer: Boolean!) {
                 disabled
                 file
             }
-            meta_title @include(if: $onServer)
-            meta_keyword @include(if: $onServer)
-            meta_description @include(if: $onServer)
         }
     }
 }