--- a/packages/pwa-buildpack/lib/WebpackTools/__tests__/__fixtures__/queries/getProductDetail.graphql
+++ b/packages/pwa-buildpack/lib/WebpackTools/__tests__/__fixtures__/queries/getProductDetail.graphql
@@ -19,6 +19,9 @@ query productDetail($urlKey: String) {
                 disabled
                 file
             }
+            meta_title
+            meta_keyword
+            meta_description
         }
     }
 }