--- a/packages/venia-ui/lib/fragments/productDetails.graphql
+++ b/packages/venia-ui/lib/fragments/productDetails.graphql
@@ -20,8 +20,6 @@ fragment ProductDetails on ProductInterface {
         disabled
         file
     }
-    meta_title @include(if: $onServer)
-    meta_keyword @include(if: $onServer)
     meta_description
     name
     price {