--- a/packages/venia-ui/lib/fragments/productDetails.graphql
+++ b/packages/venia-ui/lib/fragments/productDetails.graphql
@@ -20,6 +20,8 @@ fragment ProductDetails on ProductInterface {
         disabled
         file
     }
+    meta_title
+    meta_keyword
     meta_description
     name
     price {