--- a/packages/venia-ui/lib/queries/getCategory.graphql
+++ b/packages/venia-ui/lib/queries/getCategory.graphql
@@ -2,7 +2,6 @@ query category(
     $id: Int!
     $pageSize: Int!
     $currentPage: Int!
-    $onServer: Boolean!
     $filters: ProductAttributeFilterInput!
     $sort: ProductAttributeSortInput
 ) {
@@ -11,8 +10,6 @@ query category(
         description
         name
         product_count
-        meta_title @include(if: $onServer)
-        meta_keywords @include(if: $onServer)
         meta_description
     }
     products(
@@ -35,8 +32,6 @@ query category(
                 disabled
                 file
             }
-            meta_title @include(if: $onServer)
-            meta_keyword @include(if: $onServer)
             meta_description
             name
             price {