--- a/packages/venia-ui/lib/queries/getCategory.graphql
+++ b/packages/venia-ui/lib/queries/getCategory.graphql
@@ -11,7 +11,7 @@ query category(
         name
         product_count
         meta_title
-        meta_keyword
+        meta_keywords
         meta_description
     }
     products(