--- a/packages/venia-ui/lib/queries/getCmsPage.graphql
+++ b/packages/venia-ui/lib/queries/getCmsPage.graphql
@@ -1,12 +1,10 @@
-query getCmsPage($id: Int!, $onServer: Boolean!) {
+query getCmsPage($id: Int!) {
     cmsPage(id: $id) {
         url_key
         content
         content_heading
         title
         page_layout
-        meta_title @include(if: $onServer)
-        meta_keywords @include(if: $onServer)
         meta_description
     }
 }