--- a/packages/venia-ui/lib/queries/getCmsPage.graphql
+++ b/packages/venia-ui/lib/queries/getCmsPage.graphql
@@ -5,6 +5,8 @@ query getCmsPage($id: Int!) {
         content_heading
         title
         page_layout
+        meta_title
+        meta_keyword
         meta_description
     }
 }