--- a/packages/venia-ui/lib/queries/getProductDetail.graphql
+++ b/packages/venia-ui/lib/queries/getProductDetail.graphql
@@ -1,4 +1,4 @@
-query productDetail($urlKey: String, $onServer: Boolean!) {
+query productDetail($urlKey: String) {
     productDetail: products(filter: { url_key: { eq: $urlKey } }) {
         items {
             # Once graphql-ce/1027 is resolved, use a ProductDetails fragment
@@ -22,8 +22,6 @@ query productDetail($urlKey: String, $onServer: Boolean!) {
                 disabled
                 file
             }
-            meta_title @include(if: $onServer)
-            meta_keyword @include(if: $onServer)
             meta_description
             name
             price {