--- a/packages/venia-ui/lib/queries/getProductDetail.graphql
+++ b/packages/venia-ui/lib/queries/getProductDetail.graphql
@@ -22,6 +22,8 @@ query productDetail($urlKey: String) {
                 disabled
                 file
             }
+            meta_title
+            meta_keyword
             meta_description
             name
             price {