--- a/packages/venia-ui/package.json
+++ b/packages/venia-ui/package.json
@@ -82,6 +82,7 @@
     "apollo-cache-persist": "~0.1.1",
     "apollo-client": "2.6.4",
     "apollo-link-http": "~1.5.11",
+    "graphql": "~14.3.1",
     "braintree-web-drop-in": "~1.16.0",
     "react": "~16.9.0",
     "react-redux": "~7.1.1",