## Related Issue

Closes [PWA-644](https://jira.corp.magento.com/browse/PWA-644)

## Description

https://devdocs.magento.com/guides/v2.4/graphql/caching.html

Enables using GET for GraphQL queries. Some work-arounds were used to ensure that clients are unlikely to hit URL length limits in most environments:

- Fastly's WAF currently blocks any requests that include a match on the pattern `include(.+)`. This means that, when using Magento Cloud (Fastly) _and_ the `@include` directive in GraphQL, requests will fail. We need to work with the cloud team on this, but for the time being @sirugh / @zetlen said to just remove `@include` usage entirely (we only used it for an early ssr experiment, afaik).

- This PR makes a time/space trade-off. Apollo Client serializes queries from an AST _at runtime_, and these queries have a ton of unnecessary whitespace and punctuators. To address that, this PR takes a ~16kb hit on client-side JS, in return for significant savings on API request times. The `16kb` comes from the [`stripIgnoredCharacters` utility](https://github.com/graphql/graphql-js/blob/7b3241329e1ff49fb647b043b80568f0cf9e1a7c/src/utilities/stripIgnoredCharacters.js) in the GraphQL JavaScript reference implementation. There are other ways y'all could technically hack around this (webpack-swap the `printer` function from graphql js, and implement a better printer. I have some scratch code that works if anyone wants to take a stab at it).

- Apollo client's [HTTP link uses the typical `encodeURIComponent` function on queries](https://github.com/apollographql/apollo-client/blob/7993a778ef5a9e61bfee72b956f4c179fce691f4/src/link/http/rewriteURIForGET.ts#L11). However, thanks to "don't break the web," we can cheat a bit here since query strings allow `application/x-www-form-urlencoded` encoding. So we (hackily) re-write the URLs HTTP link generates, and use the `URLSearchParams` interface to get more efficient encoding.

## Results

**Before/After Comparisons (Note they're backwards, after is on the left)**
![image](https://user-images.githubusercontent.com/5233399/89238918-4916fd00-d5bd-11ea-92d8-87bf4ec56d0e.png)

I used the `category` query as a test, and used `https://venia.magento.com/graphql` for the API.

### Resolve Route
**Before**: 374ms
**After**: 49ms

### Navigation Query
**Before**: 428ms
**After**: 50ms

### Category Query
**Before**: 1.2s
**After**: 67ms

### Product Filters By Category
**Before**: 539ms
**After**: 76ms


In this test, the total savings was over 2 seconds for page load of a category page.

## Commits

This PR is split up into 3 separate logical commits:

- cf28ebb83fc84ad8ce806f8f8e966221027871e0 (Remove usage of `onServer`)
- 245e8d6279b3c0c272e12cdeca0f0e11d2b55cfc (Switch apollo-client to `GET`)
- e7e6fffebd1cdd78d1226c8305d9577aded83207 (Switch `resolveUnknownRoutes` to `GET`)

## Acceptance 
<!-- The people and processes this pull request needs before it is merged. -->
<!-- These fields are not required when opening the pull request, but they -->
<!-- should be populated after code review. -->
### Verification Stakeholders
Don't know!
<!-- People who must verify that this solves the attached issue. -->

### Verification Steps
<!-- Please describe in detail how a reviewer can verify your changes, -->
<!-- OR how you will demonstrate the changes to the stakeholder(s). -->
1. Go look at queries in the storefront
2. Some of them are faster!

## Screenshots / Screen Captures (if appropriate)
**Before/After Comparison of GraphQL network requests**
![image](https://user-images.githubusercontent.com/5233399/89238918-4916fd00-d5bd-11ea-92d8-87bf4ec56d0e.png)




## Checklist
<!--- Go over all the following points, and make sure you've done anything necessary -->
- [X] I have added tests to cover my changes, if necessary.
* I have updated the documentation accordingly, if necessary.