@sirugh 2020-08-04
> Fastly's WAF currently blocks any requests that include a match on the pattern include(.+). This means that, when using Magento Cloud (Fastly) and the @include directive in GraphQL, requests will fail. We need to work with the cloud team on this, but for the time being @sirugh / @zetlen said to just remove @include usage entirely (we only used it for an early ssr experiment, afaik).

@jcalcaben we should make sure to document this somewhere visible. We are removing our own usage of the directive but if anyone in the community is using it this will potentially break them.