@DrewML 2020-08-04
FYI @billygilbert is on 🔥 and already is logging a ticket for Fastly