@DrewML 2020-08-05
@jimbo 

> Right, the inline screenshots are probably too specific. A basic comparison asserting that the new approach produces a string of lesser length would suffice here.

Changed to `expect(result.length).toBeLessThan(apolloClientGETURL.length);`