@DrewML 2020-08-10
Update from @billygilbert:

he spoke with Fastly and got some example Varnish config to add exclusions for `@include`. He's going to open up a ticket to propose applying those changes for cloud customers