@DrewML 2020-08-11
There are some caveats that pwa-studio developers will want to know about caching and combining queries - opened a devdocs ticket to track: https://github.com/magento/devdocs/issues/7718